requirements.txt
asgiref==3.3.4
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
ConfigArgParse==1.4.1
fastapi==0.65.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
google-api-core==1.30.0
google-auth==1.31.0
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
locust==1.5.3
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
